moved db libraries into a database_lib folder
